Clear backend and organize it
