+++ .gitignore
@@ -160,7 +160,7 @@ yarn-error.log
 .history/*
 
 # Miscellaneous
-/.angular/cache
+/client/.angular/cache
 .sass-cache/
 /connect.lock
 /coverage
