+++ server/models/Message.js
@@ -1,4 +1,3 @@
-// models/Post.js
 const mongoose = require('mongoose');
 
 const messageSchema = new mongoose.Schema({
