--- server/sockets/chats/connections.js
@@ -1,12 +1,0 @@
-function joinRoom(socket, room) {
-  socket.join(room);
-}
-
-function leaveRoom(socket, room) {
-  socket.leave(room);
-}
-
-module.exports = {
-  joinRoom,
-  leaveRoom
-};
