+++ server/sockets/socket.js
@@ -1,0 +1,33 @@
+// websocketServer.js
+const socketIO = require('socket.io');
+
+const { sendMessage } = require('./socketMessages');
+
+const initSocketOnServer = (server) => {
+
+  const io = socketIO(server, {
+  cors: {
+    origin: '*',
+  }
+});
+
+  // New connection
+  io.on('connection', (socket) => {
+    console.log('User connected!');
+
+    // Disconnect
+    socket.on('disconnect', () => {
+      console.log('User disconnected!');
+    });
+
+    // Messages actions
+    sendMessage(io, socket);
+
+  });
+
+  return io;
+}
+
+module.exports = {
+  initSocketOnServer
+};
